\name{attributeimportanceJLI-package}
\alias{attributeimportanceJLI-package}
\alias{attributeimportanceJLI}
\docType{package}
\title{
\packageTitle{attributeimportanceJLI}
}
\description{
\packageDescription{attributeimportanceJLI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{attributeimportanceJLI}
\packageIndices{attributeimportanceJLI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{attributeimportanceJLI}

Maintainer: \packageMaintainer{attributeimportanceJLI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
