\name{attributeImportanceJLI-package}
\alias{attributeImportanceJLI-package}
\alias{attributeImportanceJLI}
\docType{package}
\title{
\packageTitle{attributeImportanceJLI}
}
\description{
\packageDescription{attributeImportanceJLI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{attributeImportanceJLI}
\packageIndices{attributeImportanceJLI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{attributeImportanceJLI}

Maintainer: \packageMaintainer{attributeImportanceJLI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
